move debug tools
